Add attribute desired_state
This is similar to google's desired_state attribute which can put a virtual machine in a terminated or running state. However, the goal is to allow designating every state which can be resumed from. To change between inactive states, the virtual machine is started and then the relevant action is called.