only set randompasswd in graphics template if it's true (#535)
Small fix in the template schemas for older OpenNebula versions. 

Older OpenNebula versions (like 5.12) interpret the `RANDOM_PASSWORD`
as truthy if its value is not empty and therefore try to generate a
password, which when using VNC will not be truncated due to another
bug that's already been fixed in the latest release.